pyproject.toml
[project]
name = "appety"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "python-dotenv>=1.0.1",
    "seleniumbase>=4.32.12",
    "slack-bolt>=1.21.2",
]
